code/requirements.txt
pyspark==3.2.0
psycopg2-binary~=2.8.5
pyyaml